For package_name, remove params.pp for module_json
This commit begins the refactoring of this module to scrap the
params.pp pattern for the experimental hiera-module_json backend.

This backend allows for module specific data to expressed in json
similar to how Hiera functions at a global level, allowing the
consumer to express their data outside of their manifests.